mi-backend/requirements.txt
# # fastapi==0.95.2
# # uvicorn==0.22.0
# # torch==2.1.0
# # torchvision==0.15.2
# # onnx==1.14.1
# # onnxruntime==1.15.0
# # scikit-learn==1.3.2
# # pillow==10.0.1
# # numpy==2.2.4
# # pandas==2.2.2
# # opencv-python==4.9.0.74
# # matplotlib==3.8.1
# # aiofiles==23.1.0
# # python-multipart==0.0.6
# # albumentations==1.3.1
# # joblib==1.3.2
# fastapi==0.95.2
# uvicorn==0.22.0
# torch==2.1.0
# torchvision==0.15.2
# onnx==1.14.1
# onnxruntime==1.15.0
# scikit-learn==1.3.2
# pillow==10.0.1
# numpy==1.26.4
# pandas==2.2.2
# opencv-python==4.10.0.84
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
onnx==1.14.1
onnxruntime==1.15.0
scikit-learn==1.3.2
pillow==10.0.1
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
medmnist==2.2.2
